+++ api/src/db/database.ts
@@ -6,6 +6,7 @@ import { ChatMessage } from '@models/message.model';
 import { TwitterPost } from '@models/twitter-posts.model';
 import { Repo } from '@models/repos.model';
 import { JobState } from '@models/job.model';
+import { GithubEvent } from '@models/github-events.model';
 
 export const AppDataSource = new DataSource({
   type: 'mysql',
@@ -14,7 +15,7 @@ export const AppDataSource = new DataSource({
   username: process.env.MYSQLUSER || 'root',
   password: process.env.DB_PASSWORD || 'password',
   database: process.env.MYSQL_DATABASE || 'repo-ranger',
-  entities: [Username, ChatMessage, TwitterPost, Repo, Event, JobState],
+  entities: [Username, ChatMessage, TwitterPost, Repo, JobState, GithubEvent],
   synchronize: true,
   logging: false,
 });