+++ api/src/models/github-events.model.ts
@@ -1,7 +1,7 @@
 import { Entity, PrimaryGeneratedColumn, Column, BaseEntity } from 'typeorm';
 
-@Entity('events')
-export class Event extends BaseEntity {
+@Entity('github_events')
+export class GithubEvent extends BaseEntity {
   @PrimaryGeneratedColumn()
   id!: number;
 
@@ -9,20 +9,23 @@ export class Event extends BaseEntity {
   username_id!: number;
 
   @Column({ type: 'int' })
-  repo_id!: number;
+  github_repo_id!: number;
 
   @Column({ type: 'varchar' })
   event_type!: string;
 
-  @Column({ type: 'datetime' })
-  created_at!: Date;
-
   @Column({ type: 'int', default: 0 })
   event_size!: number; // (that number is a multiplier for the score of this event )
 
-  @Column({ type: 'varchar', nullable: true })
+  @Column({ type: 'varchar', nullable: true, unique: true })
   github_id!: string;
 
   @Column({ type: 'text', nullable: true })
   message!: string;
+
+  @Column({ type: 'datetime' })
+  event_date!: Date;
+
+  @Column({ type: 'datetime' })
+  created_at!: Date;
 }