+++ api/src/routes/routes.ts
@@ -11,6 +11,7 @@ router.get('/single/:id', usernameController.getUsernameById);
 router.get('/search', usernameController.searchUsernamesSortedByScore);
 router.get('/example-tweet', twitterController.sendExampleTweet);
 router.get('/example-update-tweets-performance', twitterController.updateTweetsPerformance);
+router.get('/example-refresh-username/:id', usernameController.refreshUsername);
 
 router.get('/chat/messages', messageController.getLatestChatMessages);
 router.get('/chat/messages/:oldestMessageId', messageController.getOlderChatMessages);