+++ api/src/types/events.interface.ts
@@ -1,0 +1,55 @@
+export interface EventDTO {
+  username_id: number;
+  github_repo_id: number;
+  event_type: string;
+  event_size: number;
+  github_id: string;
+  message: string;
+  event_date: Date;
+}
+
+export interface GithubEvent {
+  id: string;
+  type: string;
+  actor: {
+    id: number;
+    login: string;
+    display_login: string;
+    gravatar_id: string;
+    url: string;
+    avatar_url: string;
+  };
+  repo: {
+    id: number;
+    name: string;
+    url: string;
+  };
+  payload: {
+    repository_id: number;
+    push_id: number;
+    size: number;
+    distinct_size: number;
+    ref: string;
+    head: string;
+    before: string;
+    commits: Array<{
+      sha: string;
+      author: {
+        email: string;
+        name: string;
+      };
+      message: string;
+      distinct: boolean;
+      url: string;
+    }>;
+  };
+  public: boolean;
+  created_at: string;
+  org: {
+    id: number;
+    login: string;
+    gravatar_id: string;
+    url: string;
+    avatar_url: string;
+  };
+}